Lifted restriction that extension DLLs be signed by same authority as mscript EXE.
Add restriction that alpha comparison ops like NEQ have spaces around them in order to be considered operators.

Changed > to become >>, and > to be a cmd execution statement.  Lots of test script updates!

Removed requirement that declaring a variable has to provide an initial value.  This is in keeping with JavaScript's let statement.

Layed out plans for a switch statement with %, =, and * symbols.